Remove DQM demo
