+++ requirements.txt
@@ -1,3 +1,3 @@
 imageio==2.9.0
-dwave-ocean-sdk>=4.4.0
+dwave-ocean-sdk>=4.5.0
 matplotlib
